Updated test_bundle
